v60-fix38: Stock.quantity is now Float — preserves 0.50 etc instead of truncating to 0
User bug report: 'AJ-214-276-C ei item e 0.50 stock add korlam DEWS e. but show korche na. barcode: OFF-DEWS-26060309'

Root cause:
- Stock table had quantity INTEGER in Prisma schema (and the corresponding
  DB column was INTEGER).
- When user added 0.50 via Add Stock, the value was inserted into an INTEGER
  column → SQLite truncated it to 0.
- Then my zero-stock filter (added in v60-fix37: 'quantity > 0') excluded it.
- Net effect: stock silently disappeared on the way in.

Fix:
1. prisma/schema.prisma — Stock.quantity: Int → Float
2. /api/stock/route.ts POST — parseInt(quantity) → parseFloat(quantity)
3. /api/stock/add/route.ts — removed Math.round() from set/create paths so
   the exact decimal value is stored. (Increment path already used raw qty.)
4. /api/admin/migrate-schema/route.ts — added 6-step migration (v63) that
   recreates the Stock table with quantity REAL via the SQLite
   table-recreate pattern (create new, copy with CAST AS REAL, drop old,
   rename, recreate indexes). All steps are idempotent.
5. Ran 'prisma generate' locally to refresh the Prisma client with the new
   Float type.

Post-deploy step required:
After Vercel deploys this commit, the user needs to run the migrate-schema
endpoint once on production:
  curl -X POST 'https://dfcl-inv.vercel.app/api/admin/migrate-schema?token=DFCL_RESCUE_2026'
This will convert the production Stock.quantity column from INTEGER to REAL.
Until this is run, decimal values will still be truncated on production.

After the migration runs, the user should re-add the 0.50 stock for
AJ-214-276-C (since the previous attempt was truncated to 0).

+++ src/app/api/stock/route.ts
@@ -159,8 +159,8 @@ export async function POST(request: NextRequest) {
       where: {
         itemId_entityId: { itemId, entityId },
       },
-      update: { quantity: parseInt(quantity) },
-      create: { itemId, entityId, quantity: parseInt(quantity) },
+      update: { quantity: parseFloat(quantity) },
+      create: { itemId, entityId, quantity: parseFloat(quantity) },
     });
 
     return NextResponse.json({ stock });
